// src/app/api/model/transaction-request.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TransactionRequest { 
    accountId: number;
    type: TransactionRequestTypeEnum;
    amount: number;
}
export enum TransactionRequestTypeEnum {
    Deposit = 'DEPOSIT',
    Withdrawal = 'WITHDRAWAL'
};